refactor: Simply moved account command and tests to own package
