+++ ticket-service/src/test/java/com/epam/training/ticketservice/shell/command/account/AccountCommandShellTest.java
@@ -1,4 +1,4 @@
-package com.epam.training.ticketservice.shell.command;
+package com.epam.training.ticketservice.shell.command.account;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
